+++ js/data.js
@@ -1,0 +1,34 @@
+export const PHOTO_DESCRIPTIONS = [
+  'Завтрак у моря',
+  'Кардиотренировка в спортзале',
+  'Работа из дома',
+  'Отдыхаю с друзьями',
+  'Новогоднее настроение',
+  'Купил новый телефон',
+  'Нереальные закаты',
+  'Читаю новую книгу'
+];
+
+export const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Артём',
+  'Юлия',
+  'Алина',
+  'Валерия',
+  'Мария',
+  'Денис',
+  'София',
+  'Даниил',
+  'Павел',
+  'Вадим',
+  'Галина',
+  'Елена'
+];
